Refactor test-support module and split it in two
