+++ test-support/build.gradle.kts
@@ -4,17 +4,12 @@ plugins {
   id("config-multi")
 }
 
-// assertK doesn't support windows builds yet, simply disabling the task results doesn't solve the issue
-rootProject.gradle.startParameter.excludedTaskNames.add(":test-support:compileKotlinMingwX64")
-
 kotlin {
   sourceSets {
     val commonMain by getting {
       dependencies {
-        api(libs.kotlinx.coroutines.core)
-        api(libs.kotlinx.coroutines.test)
-        api(libs.assertk.core)
-        api(libs.turbine)
+        implementation(libs.kotlinx.coroutines.core)
+        implementation(libs.kotlinx.coroutines.test)
         implementation(project(":core"))
       }
     }
